+++ x/genutil/gentx_test.go
@@ -271,9 +271,10 @@ func (suite *GenTxTestSuite) TestDeliverGenTxs() {
 				_ = suite.setAccountBalance(addr1, 50)
 				_ = suite.setAccountBalance(addr2, 1)
 
+				r := rand.New(rand.NewSource(time.Now().UnixNano()))
 				msg := banktypes.NewMsgSend(addr1, addr2, sdk.Coins{sdk.NewInt64Coin(sdk.DefaultBondDenom, 1)})
 				tx, err := helpers.GenSignedMockTx(
-					rand.New(rand.NewSource(time.Now().UnixNano())),
+					r,
 					suite.encodingConfig.TxConfig,
 					[]sdk.Msg{msg},
 					sdk.Coins{sdk.NewInt64Coin(sdk.DefaultBondDenom, 10)},